chore: move DetachUtils closer to usage (#10590)
which is only in preheat.supplier/supplier.strategy
